extract these shell helpers into their own module
