+++ lib/language_pack/base.rb
@@ -2,11 +2,14 @@ require "language_pack"
 require "pathname"
 require "yaml"
 require "digest/sha1"
+require "language_pack/shell_helpers"
 
 Encoding.default_external = Encoding::UTF_8 if defined?(Encoding)
 
 # abstract class that all the Ruby based Language Packs inherit from
 class LanguagePack::Base
+  include LanguagePack::ShellHelpers
+
   VENDOR_URL = "https://s3.amazonaws.com/heroku-buildpack-ruby"
 
   attr_reader :build_path, :cache_path
@@ -124,65 +127,6 @@ private ##################################
         else arg
       end
     end.join(" ")
-  end
-
-  # display error message and stop the build process
-  # @param [String] error message
-  def error(message)
-    Kernel.puts " !"
-    message.split("\n").each do |line|
-      Kernel.puts " !     #{line.strip}"
-    end
-    Kernel.puts " !"
-    log "exit", :error => message
-    exit 1
-  end
-
-  # run a shell comannd and pipe stderr to stdout
-  # @param [String] command to be run
-  # @return [String] output of stdout and stderr
-  def run(command)
-    %x{ #{command} 2>&1 }
-  end
-
-  # run a shell command and pipe stderr to /dev/null
-  # @param [String] command to be run
-  # @return [String] output of stdout
-  def run_stdout(command)
-    %x{ #{command} 2>/dev/null }
-  end
-
-  # run a shell command and stream the output
-  # @param [String] command to be run
-  def pipe(command)
-    output = ""
-    IO.popen(command) do |io|
-      until io.eof?
-        buffer = io.gets
-        output << buffer
-        puts buffer
-      end
-    end
-
-    output
-  end
-
-  # display a topic message
-  # (denoted by ----->)
-  # @param [String] topic message to be displayed
-  def topic(message)
-    Kernel.puts "-----> #{message}"
-    $stdout.flush
-  end
-
-  # display a message in line
-  # (indented by 6 spaces)
-  # @param [String] message to be displayed
-  def puts(message)
-    message.split("\n").each do |line|
-      super "       #{line.strip}"
-    end
-    $stdout.flush
   end
 
   # create a Pathname of the cache dir
